refactor(core): move interpolation_utils implementation into its own header
